+++ clothing-shop/src/components/category-item/category-item.component.jsx
