+++ clothing-shop/src/components/category-item/category-item.styles.scss
