+++ clothing-shop/src/components/directory/directory.component.jsx
