+++ clothing-shop/src/components/directory/directory.styles.scss
